[Refactor] useAddFolder 추가 및 모달 수정
